css-transforms-1/transform_translate_zero_prefixed.html

<!DOCTYPE html>
<html>
<head>
    <title>CSS Transforms API Test: transform translate</title>
    <link rel="author" title="Intel" href="http://www.intel.com">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#transform-property">
    <link rel="help" href="http://www.w3.org/TR/css-transforms-1/#two-d-transform-functions">
    <meta name="flags" content="dom">
    <meta name="assert" content="Check if transform supports translate(0, 0)">
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
</head>
<body>
    <div id="test"></div>
    <div id="log"></div>
    <script>
        var div = document.getElementById("test");
        test(function() {
            div.style.transform = "translate(0, 0)";
            var value = window.getComputedStyle(div).getPropertyValue("transform");

            assert_equals(value, "matrix(1, 0, 0, 1, 0, 0)", "transform should be matrix(1, 0, 0, 1, 0, 0)");
        }, "transform_translate_0_0");
    </script>
</body>
</html>
